fix: 모듈화(#6)
