+++ db/__init__.py
@@ -1,0 +1,14 @@
+
+from .chroma_client import (
+    get_chroma_client,
+    get_embedding_function,
+    get_user_latest_collection,
+    get_user_history_collection
+)
+
+__all__ = [
+    "get_chroma_client",
+    "get_embedding_function",
+    "get_user_latest_collection",
+    "get_user_history_collection"
+]
